Todo integration (#18)
* fix env errors

* add todo working

* getTodos working

* getTodos bug fixed

* todo details working

* delete todo working

* work on todo

* work on todo

* add try catch to every api call

* fixed signIn error

* added errors to all add/delete/and fetch functions

* fixed tenant code bug

* final touches

* Update tsconfig.json

* Update tsconfig.json

* Cleanups

* Sorrounding service call with try-catch, simplifying setting of properites

* Try-catch

* Update useGetLeasesAndTenants.ts

* final touches

---------

+++ src/Screens/AllProperties.tsx
@@ -13,7 +13,11 @@ function AllProperties() {
 	const appContext = useAppContext();
 
 	const handleDeleteProperty = useCallback(async (propertyId:number) => {
-		await appContext.deleteProperty(propertyId);
+		try{
+			await appContext.deleteProperty(propertyId);
+		}catch (e){
+			return;
+		}
 	}, [appContext]);
 
 	return(
@@ -25,9 +29,12 @@ function AllProperties() {
 			<SwipeListView
 				data={appContext.Properties}
 				rightOpenValue={-50}
-				renderHiddenItem={({ item }) => (
-					<TrashButton onPress={()=>handleDeleteProperty(item.PropertyId)}/>
-				)}
+				renderHiddenItem={({ item, index }) => {
+					if (appContext.Properties[index].isTenant) return null;
+					return (
+						<TrashButton onPress={() => handleDeleteProperty(item.PropertyId)} />
+					);
+				}}
 				renderItem={({ item }) => (
 					<PropertyCard property={item} />
 				)}
